Add memories: address decoding and read logic
Another read state is introduced. On the first clock the address is
decoded and a memory component is selected if it matches the region of a
memory. The second clock (new sReadSelect state) is needed for the
memory to return data at that address.

All register reads are delayed by one clock, too. An optimized design
might skip sReadSelect for registers.