tb/test_hectare: Extend OSVVM with a DPM responder
The DPM verification component (VC) models a dual-port memory (DPM) by
responding to a DPM interface that is coming from the DUT.

OSVVM transactions from the TestCtrl entity expect certain activity on
the DPM interface. For example, the GetWrite() procedure makes the VC
expect a write command with certain data on the DPM interface.

Use this to verify DPM interfaces generated from SystemRDL.

Separate processes are necessary for OSVVM transactions that initiate
requests on the AXI4-Lite (manager) side and that cause a response on
the DPM side. Both procedures are blocking.