tb/test_hectare: Test downstream AXI4-Lite interface
Use the OSVVM AXI4LiteResponder_Transactor as a model for the
downstream AXI4-Lite module that would be axi4_spi in a real design.

The AXI4 SPI adapter that is modelled expects 8 bits of data while the
DUT always provides 32 bits. The AXI4 SPI adapter should be changed to
handle that, but this is out of the scope of DesyRDL, of course.