Move all context generation to DesyListener
Currently it is the RdlFormatter class (the template engine) that parses
the information in a SystemRDL node so that it can be used in a
template. The template engine should be separate from the SystemRDL
specifics though.

With these changes, all context that the template engine needs is
created upfront.