Rename 'base' -> 'internal_offset'
The internal register address decoder in VHDL uses a std_logic_vector
with one bit per register for selecting the target register. It is
flattened like this:

0: my_reg[0]
1: my_reg[1]
2: my_reg[2]
3: my_reg[3]
4: other_reg[0]

The offset of 4 must be known to to access other_reg. The name 'base' is
confusing - 'internal_offset' should first make clear that this is
related to the internal implementation. Avoid keywords that might be
misleading, like 'array_index' might be.

Alternatives:
* first_element_index
* first_item_number