Read a list of templates, write a list of output files
The tool that uses the DesyRDL outputs (e.g. fwfwk or ModelSim) needs an
ordered list of files to be included in its project.

Read an ordered list of template inputs from a text file, or fall back
to globbing. Write a list of output files to another text file.

This way we can drop the template processing on template filenames.