Read a pre-ordered list of libraries
VHDL files that are part of the generated output, but don't require
template processing, are also a part of the list of output files. This
list is supposed to be ordered so that files can be added to the VHDL
parser in that order.

Introduce another `include.txt` where the order is already correct,
instead of just globbing for all files in the respective folder. Fall
back to globbing if it doesn't exist.