RdlFormatter: [UNTESTED] Add an "if" function
This implements a few fixed tests:

* if:gtzero := if greater than zero
* if:ltzero := if less than zero
* if:eqzero := if equal to zero
* if:nezero := if not equal to zero

The value in the template string must be a dictionary. The test is
applied to the value at the specified key. If the test succeeds, the
specified template is used for further processing, using the flattened
dictionary as arguments.

Use like this: {value:if:gtzero:key:template}