DesyListener: Fix context construction
The collected types and instance names of registers etc were cleared at
addressmap boundaries for RTL generation, but not for mapfile
generation, because there it would have to cover the complete tree.

Unify the way this information is collected and cleared at ExitAddrmap.
Construct mapfiles by appending to a single file on each addrmap.