Use the original type name if necessary
When instantiating generated code in user HDL one has to know the name
of the generated package and top entity. This was based on the final
type name of the SystemRDL-compiled node. If the type name differs from
the original type name specified in the RDL source (e.g. because a
property was modified) then this breaks the user's code.

Use the original type name for package and entity names. This is not
ideal because it breaks the case of multiple types based on the same
original addrmap.