fix: ignore address bits above self
When a DesyRDL-generated address space is integrated into a larger
design then it is probably not placed at address 0. The address decoder
should not use the full address from the bus interface, but only the
portion that it occupies itself. The upper bus structure is repsonsible
for only forwarding requests that are targeted at this instance.

Note that this is unlike PCI, for example, where each device on the bus
decodes the full address and "claims" the transaction for itself when it
is configured to decode that address.