refactor: improve hierarchy separator usage in context
It was now agreed upon with the ChimeraTK team that mapfiles can use
dots as hierarchy separators in all cases (modules, regfiles, arrays of
memories). For VHDL code, an underscore is always used.

The code that creates strings for the template engine now always uses
the filetype-specific hierarchy separator to connect individual
sub-strings.