feat: add a Dunder Version and use it in setup.cfg
The version string was already defined in setup.cfg, but there it is
only available to the build and packaging mechanism, not to the Python
runtime. That would typically be made possible by the so called "Dunder
Version" which is a `__version__` attribute.

The Python Packaging User Guide describes ways to use a single source
for both. Accessing the "Dunder Version" from setup.cfg is one.

Now you can do:

    >>> import desyrdl
    >>> print(desyrdl.__version__)
    1.0.1