fix(tmpl): avoid null ranges in records
Questa or ModelSim crashes with a SIGSEGV exception when it encounters a
record where the last entry has a null range (-1 downto 0).

Entries with null ranges are generated for fields that are defined as
not accessible to the user of that record (HW or SW side).

In those cases, and only on the records that the hardware reads from,
generate a 'dummy' entry in the record instead. The name should not
matter as no code should use it, but a record must have at least one
entry. For that reason it might also be called 'sentinel'.

Keep null ranges on records that hardware writes to. This avoids
warnings about unconnected ports.

Issue: #12421